refactor(server): peel bins routes into routers/bins.py (router-split, round-5 #51 / R5-25 PR17)
Sixth router peeled (merge-as-green off main after cache/PR16). The /api/bins
group — bin CRUD, add/remove items, copy-into-project — moves into routers/bins.py
with its 4 body models, the detail-payload builder, and the verified-copy
primitives (_unique_dest / _copy_clip_verified — critical #3: never overwrite a
same-basename clip, never delete the source).

copy_bin streams ndjson while spawning an ingest.py subprocess and shares the H3
single-flight ingest slot with the ingest routes — imported from `state` (the SAME
process-wide singleton, asserted), so the double-whisper-OOM guard still holds
across bins↔ingest. BASE_DIR (config) replaces server.ROOT for locating ingest.py.
Verbatim subprocess/streaming/GeneratorExit/slot logic. Imports auth/bins/projects/
offload/state/config/pathres — no server import.

New boundary test pins leaf/ownership/mount + name validator + slot-singleton
sharing. CRUD + copy behaviour stays covered by test_bins.py / test_bin_copy.py.
858 passed.
